+++ devices/sycl/sycl_device.h
@@ -41,6 +41,7 @@ OIDN_NAMESPACE_BEGIN
     explicit SYCLDevice(const Ref<SYCLPhysicalDevice>& physicalDevice);
 
     DeviceType getType() const override { return DeviceType::SYCL; }
+    bool isSupported() const override;
     ze_context_handle_t getZeContext() const { return zeContext; }
 
     int getInt(const std::string& name) override;